+++ inc/hooks.php
@@ -51,6 +51,7 @@ add_action( 'enqueue_block_editor_assets', __NAMESPACE__ . '\block_editor_title_
 require get_theme_file_path( 'inc/hooks/acf-blocks.php' );
 add_filter( 'block_categories_all', __NAMESPACE__ . '\acf_blocks_add_category_in_gutenberg', 10, 2 );
 add_action( 'acf/init', __NAMESPACE__ . '\acf_blocks_init' );
+add_filter( 'acf/fields/wysiwyg/toolbars', __NAMESPACE__ . '\add_custom_tinymce_toolbars' );
 
 /**
  * Form related hooks